+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog of ng-select2
 
+## V7.1.7 (2020-10-19)
+
+### Corrections
+
+- Fix case with 0 in value
+
+## V7.1.6 (2020-10-13)
+
+### Changes
+
+- Added an `id` on the select options.
+- Allow to give a specific id to an option via `Select2Option` config.
+
+### Corrections
+
+- Remove the `undefined` class wrongly added to options when `classes` is not defined on `Select2Option`.
+
 ## V7.1.5 (2020-10-06)
 
 ### Corrections